chore(cloud-foundry): remove dead ngrx action classes with CF v2 API paths
The ngrx runtime is gone since the signals refactor; these ~50
ICFAction/PaginatedAction classes carried raw CF v2 API paths and were
referenced only from entity catalog builder maps that are never invoked.

- delete 8 all-dead action files; trim 6 mixed files to their live classes
- clean every builder-map entry referencing a removed class; emptied
  builder interfaces become type aliases of OrchestratedActionBuilders
- remove the caller-less createGetApplicationAction factory and its
  orphaned imports from application.service.ts
- move IUserProvidedServiceInstanceData(+Credentials) to
  cf-api-svc.types.ts beside IUserProvidedServiceInstance
- drop dead public_api exports; resolve all three FLAG (A6) markers

Closes #5726
