+++ src/frontend/packages/cloud-foundry/src/actions/application.actions.ts
@@ -1,32 +1,16 @@
 import { HttpParams, HttpRequest } from '@angular/common/http';
 
-import { pick } from '../../../store/src/helpers/reducer.helper';
-import { ActionMergeFunction } from '../../../store/src/types/api.types';
 import { PaginatedAction, PaginationParam } from '../../../store/src/types/pagination.types';
 import { ICFAction } from '../../../store/src/types/request.types';
-import { IApp } from '../cf-api.types';
 import { cfEntityFactory } from '../cf-entity-factory';
 import { applicationEntityType, appStatsEntityType } from '../cf-entity-types';
 import { CF_ENDPOINT_TYPE } from '../cf-types';
 import { createEntityRelationPaginationKey, EntityInlineParentAction } from '../entity-relations/entity-relations.types';
-import { AppMetadataTypes } from './app-metadata.actions';
 import { CFStartAction } from './cf-action.types';
 
 const GET_ALL = '[Application] Get all';
 const GET_ALL_SUCCESS = '[Application] Get all success';
 const GET_ALL_FAILED = '[Application] Get all failed';
-
-const GET = '[Application] Get one';
-const GET_SUCCESS = '[Application] Get one success';
-const GET_FAILED = '[Application] Get one failed';
-
-const CREATE = '[Application] Create';
-const CREATE_SUCCESS = '[Application] Create success';
-const CREATE_FAILED = '[Application] Create failed';
-
-export const CF_APP_UPDATE = '[Application] Update';
-export const CF_APP_UPDATE_SUCCESS = '[Application] Update success';
-export const CF_APP_UPDATE_FAILED = '[Application] Update failed';
 
 const DELETE = '[Application] Delete';
 const DELETE_SUCCESS = '[Application] Delete success';
@@ -35,10 +19,6 @@ const DELETE_FAILED = '[Application] Delete failed';
 const DELETE_INSTANCE = '[Application Instance] Delete';
 const DELETE_INSTANCE_SUCCESS = '[Application Instance] Delete success';
 const DELETE_INSTANCE_FAILED = '[Application Instance] Delete failed';
-
-const RESTAGE = '[Application] Restage';
-const RESTAGE_SUCCESS = '[Application] Restage success';
-const RESTAGE_FAILED = '[Application] Restage failed';
 
 const applicationEntitySchema = cfEntityFactory(applicationEntityType);
 
@@ -66,96 +46,6 @@ export class GetAllApplications extends CFStartAction implements PaginatedAction
   };
   flattenPagination = true;
   flattenPaginationMax = true;
-}
-
-export class GetApplication extends CFStartAction implements ICFAction, EntityInlineParentAction {
-  constructor(public guid: string, public endpointGuid: string, public includeRelations: string[] = [], public populateMissing = true) {
-    super();
-    this.options = new HttpRequest(
-      'GET',
-      `apps/${guid}`
-    );
-  }
-  actions = [GET, GET_SUCCESS, GET_FAILED];
-  entity = [applicationEntitySchema];
-  entityType = applicationEntityType;
-  options: HttpRequest<any>;
-}
-
-export class CreateNewApplication extends CFStartAction implements ICFAction {
-  constructor(
-    public guid: string,
-    public endpointGuid: string,
-    application: IApp
-  ) {
-    super();
-    // FLAG (A6): URL not swapped to /pp/v1/cf/apps/${endpointGuid} because
-    // backend createNativeApp expects v3 capi.AppCreateRequest body
-    // ({name, relationships:{space:{data:{guid}}}, ...}); current body is
-    // legacy V2 {name, space_guid}. Body-shape transform required before swap.
-    this.options = new HttpRequest(
-      'POST',
-      'apps',
-      {
-        name: application.name,
-        space_guid: application.space_guid
-      }
-    );
-  }
-  actions = [CREATE, CREATE_SUCCESS, CREATE_FAILED];
-  entity = [applicationEntitySchema];
-  entityType = applicationEntityType;
-  options: HttpRequest<any>;
-}
-
-export interface UpdateApplication {
-  name?: string;
-  instances?: number;
-  memory?: number;
-  enable_ssh?: boolean;
-  environment_json?: any;
-  state?: string;
-}
-
-export class UpdateExistingApplication extends CFStartAction implements ICFAction {
-  static updateKey = 'Updating-Existing-Application';
-
-  /**
-   * Creates an instance of UpdateExistingApplication.
-   * @param newApplication Sparsely populated application containing updated settings
-   * @param [existingApplication] Existing application. Used in a few specific cases
-   * @param [updateEntities] List of metadata calls to make if we successfully update the application
-   */
-  constructor(
-    public guid: string,
-    public endpointGuid: string,
-    public newApplication: UpdateApplication,
-    public existingApplication?: IApp,
-    public updateEntities?: AppMetadataTypes[]
-  ) {
-    super();
-    this.options = new HttpRequest(
-      'PUT',
-      `apps/${guid}`,
-      newApplication
-    );
-  }
-  actions = [CF_APP_UPDATE, CF_APP_UPDATE_SUCCESS, CF_APP_UPDATE_FAILED];
-  entity = [applicationEntitySchema];
-  entityType = applicationEntityType;
-  options: HttpRequest<any>;
-  updatingKey = UpdateExistingApplication.updateKey;
-  entityMerge: ActionMergeFunction = (oldEntities: any, newEntities: any) => {
-    const keepFromOld = pick(
-      oldEntities[applicationEntityType][this.guid].entity,
-      Object.keys(applicationEntitySchema.schema)
-    );
-    newEntities[applicationEntityType][this.guid].entity = {
-      ...newEntities[applicationEntityType][this.guid].entity,
-      ...keepFromOld
-    };
-    return newEntities;
-  };
 }
 
 export class DeleteApplication extends CFStartAction implements ICFAction {
@@ -202,20 +92,4 @@ export class DeleteApplicationInstance extends CFStartAction
   entityType = appStatsEntityType;
   removeEntityOnDelete = true;
   options: HttpRequest<any>;
-}
-
-export class RestageApplication extends CFStartAction implements ICFAction {
-  constructor(public guid: string, public endpointGuid: string) {
-    super();
-    this.options = new HttpRequest(
-      'POST',
-      `apps/${guid}/restage`,
-      null,
-    );
-  }
-  actions = [RESTAGE, RESTAGE_SUCCESS, RESTAGE_FAILED];
-  entity = [applicationEntitySchema];
-  entityType = applicationEntityType;
-  options: HttpRequest<any>;
-  updatingKey = 'restaging';
 }
