+++ src/frontend/packages/cloud-foundry/src/actions/space.actions.ts
@@ -1,68 +1,19 @@
 import { HttpParams, HttpRequest } from '@angular/common/http';
 
-import { getActions } from '../../../store/src/actions/action.helper';
 import { PaginatedAction } from '../../../store/src/types/pagination.types';
 import { ICFAction } from '../../../store/src/types/request.types';
-import { IUpdateSpace } from '../cf-api.types';
 import { cfEntityFactory } from '../cf-entity-factory';
-import {
-  applicationEntityType,
-  serviceEntityType,
-  serviceInstancesEntityType,
-  serviceInstancesWithSpaceEntityType,
-  servicePlanEntityType,
-  spaceEntityType,
-  spaceWithOrgEntityType,
-} from '../cf-entity-types';
-import { CFEntityConfig } from '../cf-types';
-import {
-  createEntityRelationKey,
-  EntityInlineChildAction,
-  EntityInlineParentAction,
-} from '../entity-relations/entity-relations.types';
+import { spaceEntityType, spaceWithOrgEntityType } from '../cf-entity-types';
+import { EntityInlineParentAction } from '../entity-relations/entity-relations.types';
 import { CFStartAction } from './cf-action.types';
-import { GetAllOrgUsers } from './organization.actions';
-import { getServiceInstanceRelations } from './service-instances.actions';
 
 export const GET_SPACES = '[Space] Get all';
 export const GET_SPACES_SUCCESS = '[Space] Get all success';
 export const GET_SPACES_FAILED = '[Space] Get all failed';
 
-export const GET_SPACE = '[Space] Get one';
-export const GET_SPACE_SUCCESS = '[Space] Get one success';
-export const GET_SPACE_FAILED = '[Space] Get one failed';
-
-export const CREATE_SPACE = '[Space] Create';
-export const CREATE_SPACE_SUCCESS = '[Space] Create Success';
-export const CREATE_SPACE_FAILED = '[Space] Create Failed';
-
 export const DELETE_SPACE = '[Space] Delete';
 export const DELETE_SPACE_SUCCESS = '[Space] Delete Success';
 export const DELETE_SPACE_FAILED = '[Space] Delete Failed';
-
-export class GetSpace extends CFStartAction implements ICFAction, EntityInlineParentAction {
-  constructor(
-    public guid: string,
-    public endpointGuid: string,
-    public includeRelations: string[] = [],
-    public populateMissing = true
-  ) {
-    super();
-    this.options = new HttpRequest(
-      'GET',
-      `spaces/${guid}`
-    );
-  }
-  actions = [
-    GET_SPACE,
-    GET_SPACE_SUCCESS,
-    GET_SPACE_FAILED
-  ];
-  entity = [cfEntityFactory(spaceEntityType)];
-  schemaKey = ''; // Required by builder
-  entityType = spaceEntityType;
-  options: HttpRequest<any>;
-}
 
 export class GetAllSpaces extends CFStartAction implements PaginatedAction, EntityInlineParentAction {
   constructor(
@@ -89,37 +40,6 @@ export class GetAllSpaces extends CFStartAction implements PaginatedAction, Enti
     'order-by': 'name'
   };
 }
-
-export class GetAllAppsInSpace extends CFStartAction implements PaginatedAction, EntityInlineParentAction, EntityInlineChildAction {
-  constructor(
-    public endpointGuid: string,
-    public spaceGuid: string,
-    public paginationKey: string,
-    public includeRelations: string[] = [],
-    public populateMissing = true,
-    public flattenPagination = true
-  ) {
-    super();
-    this.options = new HttpRequest(
-      'GET',
-      `spaces/${spaceGuid}/apps`
-    );
-    this.parentGuid = spaceGuid;
-  }
-  actions = getActions('Spaces', 'Get Apps');
-  entity = [cfEntityFactory(applicationEntityType)];
-  entityType = applicationEntityType;
-  options: HttpRequest<any>;
-  initialParams = {
-    page: 1,
-    'results-per-page': 100,
-    'order-direction': 'desc',
-    'order-direction-field': 'creation',
-  };
-  parentGuid: string;
-  parentEntityConfig = new CFEntityConfig(spaceEntityType);
-}
-
 
 export abstract class BaseSpaceAction extends CFStartAction implements ICFAction {
   actions!: string[];
@@ -150,127 +70,4 @@ export class DeleteSpace extends BaseSpaceAction {
   }
   actions = [DELETE_SPACE, DELETE_SPACE_SUCCESS, DELETE_SPACE_FAILED];
   removeEntityOnDelete = true;
-}
-
-export class CreateSpace extends BaseSpaceAction {
-  // FLAG (A6): URL not swapped to /pp/v1/cf/spaces/${endpointGuid} because
-  // backend createNativeSpace expects v3 capi.SpaceCreateRequest body
-  // ({name, relationships:{organization:{data:{guid}}}}); current body is
-  // legacy V2 IUpdateSpace ({name, organization_guid, ...}). Body-shape
-  // transform required before swap.
-  constructor(public endpointGuid: string, orgGuid: string, createSpace: IUpdateSpace, key = `${orgGuid}-${createSpace.name}`) {
-    super(key, orgGuid, endpointGuid);
-    this.options = new HttpRequest(
-      'POST',
-      'spaces',
-      createSpace
-    );
-  }
-  actions = [CREATE_SPACE, CREATE_SPACE_SUCCESS, CREATE_SPACE_FAILED];
-}
-export class UpdateSpace extends CFStartAction implements ICFAction {
-
-  public static UpdateExistingSpace = 'Updating-Existing-Space';
-  // FLAG (A6): URL/method not swapped to PATCH /pp/v1/cf/spaces/${endpointGuid}/${guid}
-  // because backend updateNativeSpace expects v3 capi.SpaceUpdateRequest body
-  // ({name?, metadata?}); current body is V2 IUpdateSpace
-  // (name, organization_guid, *_guids, allow_ssh, ...). Body-shape transform
-  // and PUT→PATCH switch required before swap.
-  constructor(public guid: string, public endpointGuid: string, updateSpace: IUpdateSpace) {
-    super();
-    this.options = new HttpRequest(
-      'PUT',
-      `spaces/${guid}`,
-      updateSpace
-    );
-  }
-  actions = getActions('Spaces', 'Update Space');
-  entity = [cfEntityFactory(spaceEntityType)];
-  entityType = spaceEntityType;
-  options: HttpRequest<any>;
-  updatingKey = UpdateSpace.UpdateExistingSpace;
-}
-
-export class GetAllSpaceUsers extends GetAllOrgUsers {
-  constructor(
-    public guid: string,
-    public paginationKey: string,
-    public endpointGuid: string,
-    public isAdmin: boolean,
-    includeRelations?: string[]) {
-    super(guid, paginationKey, endpointGuid, isAdmin, includeRelations);
-    this.options = new HttpRequest(
-      'GET',
-      `spaces/${guid}/user_roles`
-    );
-  }
-  actions = getActions('Spaces', 'List all user roles');
-}
-
-
-export class GetAllServicesForSpace extends CFStartAction implements PaginatedAction, EntityInlineParentAction {
-  constructor(
-    public paginationKey: string,
-    public endpointGuid: string = '',
-    public spaceGuid: string,
-    public includeRelations: string[] = [
-      createEntityRelationKey(serviceEntityType, servicePlanEntityType)
-    ],
-    public populateMissing = true
-  ) {
-    super();
-    this.options = new HttpRequest(
-      'GET',
-      `spaces/${spaceGuid}/services`
-    );
-  }
-  actions = getActions('Space', 'Get all Services');
-  entity = cfEntityFactory(serviceEntityType);
-  entityType = serviceEntityType;
-  options: HttpRequest<any>;
-  initialParams = {
-    page: 1,
-    'results-per-page': 100,
-    'order-direction': 'desc',
-    'order-direction-field': 'creation',
-  };
-  flattenPagination = true;
-}
-
-
-export class GetServiceInstancesForSpace
-  extends CFStartAction implements PaginatedAction, EntityInlineParentAction, EntityInlineChildAction {
-  constructor(
-    public spaceGuid: string,
-    public endpointGuid: string,
-    public paginationKey: string,
-    public q?: string[],
-    public includeRelations: string[] = getServiceInstanceRelations,
-    public populateMissing = true,
-    public flattenPagination = true
-  ) {
-    super();
-    this.options = new HttpRequest(
-      'GET',
-      `spaces/${spaceGuid}/service_instances`
-    );
-    if (q) {
-      this.initialParams.q = q;
-    }
-    this.parentGuid = spaceGuid;
-  }
-  actions = getActions('Space', 'Get all service instances');
-  entity = [cfEntityFactory(serviceInstancesWithSpaceEntityType)];
-  schemaKey = serviceInstancesWithSpaceEntityType;
-  entityType = serviceInstancesEntityType;
-  options: HttpRequest<any>;
-  initialParams = {
-    page: 1,
-    'results-per-page': 100,
-    'order-direction': 'desc',
-    'order-direction-field': 'creation',
-    q: [] as string[]
-  };
-  parentGuid: string;
-  parentEntityConfig = new CFEntityConfig(spaceEntityType);
 }
