+++ src/frontend/packages/cloud-foundry/src/entity-action-builders/application-env-var.action-builders.ts
@@ -1,10 +1,6 @@
 import { OrchestratedActionBuilders } from '../../../store/src/entity-catalog/action-orchestrator/action-orchestrator';
-import { GetAppEnvVarsAction } from '../actions/app-metadata.actions';
 
-export interface AppEnvVarActionBuilders extends OrchestratedActionBuilders {
-  getMultiple: (appGuid: string, endpointGuid: string) => GetAppEnvVarsAction;
-}
+export type AppEnvVarActionBuilders = OrchestratedActionBuilders;
 
 export const appEnvVarActionBuilders: AppEnvVarActionBuilders = {
-  getMultiple: (appGuid: string, endpointGuid: string) => new GetAppEnvVarsAction(appGuid, endpointGuid),
 };
