+++ src/frontend/packages/cloud-foundry/src/entity-action-builders/application-stats.action-builders.ts
@@ -1,10 +1,6 @@
 import { OrchestratedActionBuilders } from '../../../store/src/entity-catalog/action-orchestrator/action-orchestrator';
-import { GetAppStatsAction } from '../actions/app-metadata.actions';
 
-export interface AppStatsActionBuilders extends OrchestratedActionBuilders {
-  getMultiple: (appGuid: string, endpointGuid: string) => GetAppStatsAction;
-}
+export type AppStatsActionBuilders = OrchestratedActionBuilders;
 
 export const appStatsActionBuilders: AppStatsActionBuilders = {
-  getMultiple: (appGuid: string, endpointGuid: string) => new GetAppStatsAction(appGuid, endpointGuid)
 };
