+++ src/frontend/packages/cloud-foundry/src/entity-action-builders/application-summary.action-builders.ts
@@ -1,10 +1,6 @@
 import { OrchestratedActionBuilders } from '../../../store/src/entity-catalog/action-orchestrator/action-orchestrator';
-import { GetAppSummaryAction } from '../actions/app-metadata.actions';
 
-export interface AppSummaryActionBuilders extends OrchestratedActionBuilders {
-  get: (appGuid: string, endpointGuid: string) => GetAppSummaryAction;
-}
+export type AppSummaryActionBuilders = OrchestratedActionBuilders;
 
 export const appSummaryActionBuilders: AppSummaryActionBuilders = {
-  get: (appGuid: string, endpointGuid: string) => new GetAppSummaryAction(appGuid, endpointGuid)
 };
