+++ src/frontend/packages/cloud-foundry/src/entity-action-builders/quota-definition.action-builders.ts
@@ -1,12 +1,6 @@
 import { OrchestratedActionBuilders } from '../../../store/src/entity-catalog/action-orchestrator/action-orchestrator';
-import {
-  CreateQuotaDefinition,
-  GetQuotaDefinition,
-  GetQuotaDefinitions,
-  UpdateQuotaDefinition,
-} from '../actions/quota-definitions.actions';
+import { GetQuotaDefinition, GetQuotaDefinitions } from '../actions/quota-definitions.actions';
 import { CFBasePipelineRequestActionMeta } from '../cf-entity-generator';
-import { QuotaFormValues } from '../features/cf/quota-definition-form/quota-definition-form.component';
 
 export interface QuotaDefinitionActionBuilder extends OrchestratedActionBuilders {
   get: (
@@ -14,16 +8,6 @@ export interface QuotaDefinitionActionBuilder extends OrchestratedActionBuilders
     endpointGuid: string,
     meta?: CFBasePipelineRequestActionMeta
   ) => GetQuotaDefinition;
-  create: (
-    id: string,
-    endpointGuid: string,
-    createQuota: QuotaFormValues
-  ) => CreateQuotaDefinition;
-  update: (
-    guid: string,
-    endpointGuid: string,
-    updateQuota: QuotaFormValues
-  ) => UpdateQuotaDefinition;
   getMultiple: (
     paginationKey: string,
     endpointGuid: string,
@@ -37,20 +21,9 @@ export const quotaDefinitionActionBuilder: QuotaDefinitionActionBuilder = {
     endpointGuid: string,
     { includeRelations, populateMissing }: CFBasePipelineRequestActionMeta = {}
   ) => new GetQuotaDefinition(guid, endpointGuid, includeRelations, populateMissing),
-  create: (
-    id: string,
-    endpointGuid: string,
-    createQuota: QuotaFormValues
-  ) => new CreateQuotaDefinition(id, endpointGuid, createQuota),
-  update: (
-    guid: string,
-    endpointGuid: string,
-    updateQuota: QuotaFormValues
-  ) => new UpdateQuotaDefinition(guid, endpointGuid, updateQuota),
   getMultiple: (
     paginationKey: string,
     endpointGuid: string,
     { includeRelations, populateMissing }: CFBasePipelineRequestActionMeta = {}
   ) => new GetQuotaDefinitions(paginationKey, endpointGuid, includeRelations, populateMissing),
 };
-
